requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
python-multipart
customtkinter
